Picked up new makefile

+++ Makefile
@@ -59,18 +59,13 @@ gg: build
 	clear && ./$(OUT) && rm -rf *.o *~ $(OUT) *~
 
 test: $(TOBJ) $(OBJFILES)
-	$(CPPC) $^ -o $(TEST) $(CPPFLAGS) $(TESTFLAGS)
+	$(CPPC) $^ -o $(TEST) $(CPPFLAGS) $(TESTFLAGS) && ./$(TEST) && lcov -c -d . -o cov.info && genhtml -o html cov.info && rm cov.info && open html/index.html && rm -rf $(OBJDIR)/*.gcda
 
-.PHONY: clean libtest cov swipe
+.PHONY: clean libtest
 
 libtest:
 	$(CPPC) -isystem $(GTESTDIR)/include -I$(GTESTDIR) -pthread -c $(GTESTDIR)/src/gtest-all.cc -o $(OBJDIR)/gtest-all.o
 	ar -rv $(LIBDIR)/libgtest.a $(OBJDIR)/gtest-all.o
 
 clean:
-	rm -rf $(OBJDIR)/* *~ $(foreach exec, $(OUT) $(TEST), ./$(exec)) $(DEPSDIR)/deps.make
-cov:
-	lcov -c -d . -o cov.info && genhtml -o html cov.info && rm cov.info
-
-swipe:
-	rm $(OBJDIR)/*.gcda
+	@ rm -rf $(OBJDIR)/* *~ $(foreach exec, $(OUT) $(TEST), ./$(exec)) $(DEPSDIR)/deps.make
